+++ src/integrations/GA4/browser.js
@@ -24,8 +24,7 @@ export default class GA4 {
     this.analytics = analytics;
     this.measurementId = config.measurementId;
     this.capturePageView = config.capturePageView || 'rs';
-    this.addSendToParameter = config.addSendToParameter || false;
-    this.addSendToParameter = config.addSendToParameter || false;
+    this.overrideSessionId = config.overrideSessionId || false;
     this.extendPageViewParams = config.extendPageViewParams || false;
     this.isHybridModeEnabled = config.useNativeSDKToSend === false || false;
   }
@@ -51,6 +50,9 @@ export default class GA4 {
 
     gtagParameterObject.cookie_prefix = 'rs';
     gtagParameterObject.client_id = this.analytics.anonymousId;
+    if (this.isHybridModeEnabled && this.overrideSessionId) {
+      gtagParameterObject.session_id = this.analytics.uSession.sessionInfo.id;
+    }
     gtagParameterObject.debug_mode = true;
 
     if (Object.keys(gtagParameterObject).length === 0) {
@@ -66,10 +68,6 @@ export default class GA4 {
     window.gtag('get', this.measurementId, 'session_id', (sessionId) => {
       this.sessionId = sessionId;
     });
-
-    // To disable debug mode, exclude the 'debug_mode' parameter;
-    // Setting the parameter to false doesn't disable debug mode.
-    // Ref: https://support.google.com/analytics/answer/7201382?hl=en#zippy=%2Cglobal-site-tag-websites
 
     ScriptLoader(
       'google-analytics 4',
@@ -156,9 +154,7 @@ export default class GA4 {
       throw Error('Payload must have required parameters..');
     }
     const params = { ...parameters };
-    if (this.addSendToParameter) {
-      params.send_to = this.measurementId;
-    }
+    params.send_to = this.measurementId;
     if (this.analytics.userId) {
       params.user_id = this.analytics.userId;
     }
@@ -243,9 +239,7 @@ export default class GA4 {
       if (!pageProps) return;
       pageProps = flattenJsonPayload(pageProps);
       const properties = { ...getPageViewProperty(pageProps) };
-      if (this.addSendToParameter) {
-        properties.send_to = this.measurementId;
-      }
+      properties.send_to = this.measurementId;
       if (this.analytics.userId) {
         properties.user_id = this.analytics.userId;
       }
@@ -272,11 +266,12 @@ export default class GA4 {
     if (this.analytics.userId) {
       traits.user_id = this.analytics.userId;
     }
+    traits.send_to = this.measurementId;
 
     getDestinationEventName(rudderElement.message.type).forEach((events) => {
       this.sendGAEvent(events.dest, {
         group_id: groupId,
-        ...(traits || {}),
+        ...traits,
       });
     });
   }